// src/proto-ts/meta_transaction.pb.ts
export interface LiquiditySource {
  name: string;
  proportion: BigNumberJs;
  intermediateToken?: string | null | undefined;
  hops: string[];
}

export interface Quote {
  inputTokenAddress: string;
  inputAmount: BigNumberJs;
  outputTokenAddress: string;
  outputAmount: BigNumberJs;
  price: BigNumberJs;
  estimatedPriceImpact?: BigNumberJs | null | undefined;
  inputTokenToNativeRate: BigNumberJs;
  outputTokenToNativeRate: BigNumberJs;
  chainId: number;
  gas: BigNumberJs;
  estimatedGas: BigNumberJs;
  gasPrice: BigNumberJs;
  value: BigNumberJs;
  protocolFee: BigNumberJs;
  minimumProtocolFee: BigNumberJs;
  allowanceTarget: string;
  liquiditySources: LiquiditySource[];
}

/**
 **
 * MetaTransaction
 * https://github.com/0xProject/protocol/blob/development/packages/protocol-utils/src/meta_transactions.ts#L39
 */
export interface MetaTransaction {
  signerAddress: string;
  senderAddress: string;
  minGasPrice: BigNumberJs;
  maxGasPrice: BigNumberJs;
  expirationTimeSeconds: BigNumberJs;
  salt: BigNumberJs;
  callData: string;
  value: BigNumberJs;
  feeTokenAddress: string;
  feeAmount: BigNumberJs;
  chainId: number;
  verifyingContract: string;
}

export interface GetQuoteRequest {
  inputTokenAddress: string;
  outputTokenAddress: string;
  inputAmount?: BigNumberJs | null | undefined;
  outputAmount?: BigNumberJs | null | undefined;
  chainId: number;
  takerAddress: string;
  integratorId: string;
  affiliateAddress?: string | null | undefined;
  slippagePercentage?: BigNumberJs | null | undefined;
  /**
   * ID to use for the quote report `decodedUniqueId`
   */
  quoteUniqueId?: string | null | undefined;
}

export interface GetQuoteResponse {
  quote: Quote;
  metaTransaction: MetaTransaction;
}
// ...
export interface MetaTransactionService<Context = unknown> {
  GetQuote: (
    getQuoteRequest: GetQuoteRequest,
    context: Context
  ) => Promise<GetQuoteResponse> | GetQuoteResponse;
}
// ...
export const MetaTransaction = {
  /**
   * Serializes MetaTransaction to protobuf.
   */
  encode: function (msg: Partial<MetaTransaction>): Uint8Array {
    return MetaTransaction._writeMessage(
      msg,
      new BinaryWriter()
    ).getResultBuffer();
  },

  /**
   * Deserializes MetaTransaction from protobuf.
   */
  decode: function (bytes: ByteSource): MetaTransaction {
    return MetaTransaction._readMessage(
      MetaTransaction.initialize(),
      new BinaryReader(bytes)
    );
  },

  /**
   * Initializes MetaTransaction with all fields set to their default value.
   */
  initialize: function (): MetaTransaction {
    return {
      signerAddress: "",
      senderAddress: "",
      minGasPrice: BigNumberJs.initialize(),
      maxGasPrice: BigNumberJs.initialize(),
      expirationTimeSeconds: BigNumberJs.initialize(),
      salt: BigNumberJs.initialize(),
      callData: "",
      value: BigNumberJs.initialize(),
      feeTokenAddress: "",
      feeAmount: BigNumberJs.initialize(),
      chainId: 0,
      verifyingContract: "",
    };
  },

  /**
   * @private
   */
  _writeMessage: function (
    msg: Partial<MetaTransaction>,
    writer: BinaryWriter
  ): BinaryWriter {
    if (msg.signerAddress) {
      writer.writeString(1, msg.signerAddress);
    }
    if (msg.senderAddress) {
      writer.writeString(2, msg.senderAddress);
    }
    if (msg.minGasPrice) {
      writer.writeMessage(3, msg.minGasPrice, BigNumberJs._writeMessage);
    }
    if (msg.maxGasPrice) {
      writer.writeMessage(4, msg.maxGasPrice, BigNumberJs._writeMessage);
    }
    if (msg.expirationTimeSeconds) {
      writer.writeMessage(
        5,
        msg.expirationTimeSeconds,
        BigNumberJs._writeMessage
      );
    }
    if (msg.salt) {
      writer.writeMessage(6, msg.salt, BigNumberJs._writeMessage);
    }
    if (msg.callData) {
      writer.writeString(7, msg.callData);
    }
    if (msg.value) {
      writer.writeMessage(8, msg.value, BigNumberJs._writeMessage);
    }
    if (msg.feeTokenAddress) {
      writer.writeString(9, msg.feeTokenAddress);
    }
    if (msg.feeAmount) {
      writer.writeMessage(10, msg.feeAmount, BigNumberJs._writeMessage);
    }
    if (msg.chainId) {
      writer.writeInt32(11, msg.chainId);
    }
    if (msg.verifyingContract) {
      writer.writeString(12, msg.verifyingContract);
    }
    return writer;
  },

  /**
   * @private
   */
  _readMessage: function (
    msg: MetaTransaction,
    reader: BinaryReader
  ): MetaTransaction {
    while (reader.nextField()) {
      const field = reader.getFieldNumber();
      switch (field) {
        case 1: {
          msg.signerAddress = reader.readString();
          break;
        }
        case 2: {
          msg.senderAddress = reader.readString();
          break;
        }
        case 3: {
          reader.readMessage(msg.minGasPrice, BigNumberJs._readMessage);
          break;
        }
        case 4: {
          reader.readMessage(msg.maxGasPrice, BigNumberJs._readMessage);
          break;
        }
        case 5: {
          reader.readMessage(
            msg.expirationTimeSeconds,
            BigNumberJs._readMessage
          );
          break;
        }
        case 6: {
          reader.readMessage(msg.salt, BigNumberJs._readMessage);
          break;
        }
        case 7: {
          msg.callData = reader.readString();
          break;
        }
        case 8: {
          reader.readMessage(msg.value, BigNumberJs._readMessage);
          break;
        }
        case 9: {
          msg.feeTokenAddress = reader.readString();
          break;
        }
        case 10: {
          reader.readMessage(msg.feeAmount, BigNumberJs._readMessage);
          break;
        }
        case 11: {
          msg.chainId = reader.readInt32();
          break;
        }
        case 12: {
          msg.verifyingContract = reader.readString();
          break;
        }
        default: {
          reader.skipField();
          break;
        }
      }
    }
    return msg;
  },
};

export const GetQuoteRequest = {
  /**
   * Serializes GetQuoteRequest to protobuf.
   */
  encode: function (msg: Partial<GetQuoteRequest>): Uint8Array {
    return GetQuoteRequest._writeMessage(
      msg,
      new BinaryWriter()
    ).getResultBuffer();
  },

  /**
   * Deserializes GetQuoteRequest from protobuf.
   */
  decode: function (bytes: ByteSource): GetQuoteRequest {
    return GetQuoteRequest._readMessage(
      GetQuoteRequest.initialize(),
      new BinaryReader(bytes)
    );
  },

  /**
   * Initializes GetQuoteRequest with all fields set to their default value.
   */
  initialize: function (): GetQuoteRequest {
    return {
      inputTokenAddress: "",
      outputTokenAddress: "",
      inputAmount: undefined,
      outputAmount: undefined,
      chainId: 0,
      takerAddress: "",
      integratorId: "",
      affiliateAddress: undefined,
      slippagePercentage: undefined,
      quoteUniqueId: undefined,
    };
  },

  /**
   * @private
   */
  _writeMessage: function (
    msg: Partial<GetQuoteRequest>,
    writer: BinaryWriter
  ): BinaryWriter {
    if (msg.inputTokenAddress) {
      writer.writeString(1, msg.inputTokenAddress);
    }
    if (msg.outputTokenAddress) {
      writer.writeString(2, msg.outputTokenAddress);
    }
    if (msg.inputAmount != undefined) {
      writer.writeMessage(3, msg.inputAmount, BigNumberJs._writeMessage);
    }
    if (msg.outputAmount != undefined) {
      writer.writeMessage(4, msg.outputAmount, BigNumberJs._writeMessage);
    }
    if (msg.chainId) {
      writer.writeInt32(5, msg.chainId);
    }
    if (msg.takerAddress) {
      writer.writeString(6, msg.takerAddress);
    }
    if (msg.integratorId) {
      writer.writeString(7, msg.integratorId);
    }
    if (msg.affiliateAddress != undefined) {
      writer.writeString(8, msg.affiliateAddress);
    }
    if (msg.slippagePercentage != undefined) {
      writer.writeMessage(9, msg.slippagePercentage, BigNumberJs._writeMessage);
    }
    if (msg.quoteUniqueId != undefined) {
      writer.writeString(10, msg.quoteUniqueId);
    }
    return writer;
  },

  /**
   * @private
   */
  _readMessage: function (
    msg: GetQuoteRequest,
    reader: BinaryReader
  ): GetQuoteRequest {
    while (reader.nextField()) {
      const field = reader.getFieldNumber();
      switch (field) {
        case 1: {
          msg.inputTokenAddress = reader.readString();
          break;
        }
        case 2: {
          msg.outputTokenAddress = reader.readString();
          break;
        }
        case 3: {
          msg.inputAmount = BigNumberJs.initialize();
          reader.readMessage(msg.inputAmount, BigNumberJs._readMessage);
          break;
        }
        case 4: {
          msg.outputAmount = BigNumberJs.initialize();
          reader.readMessage(msg.outputAmount, BigNumberJs._readMessage);
          break;
        }
        case 5: {
          msg.chainId = reader.readInt32();
          break;
        }
        case 6: {
          msg.takerAddress = reader.readString();
          break;
        }
        case 7: {
          msg.integratorId = reader.readString();
          break;
        }
        case 8: {
          msg.affiliateAddress = reader.readString();
          break;
        }
        case 9: {
          msg.slippagePercentage = BigNumberJs.initialize();
          reader.readMessage(msg.slippagePercentage, BigNumberJs._readMessage);
          break;
        }
        case 10: {
          msg.quoteUniqueId = reader.readString();
          break;
        }
        default: {
          reader.skipField();
          break;
        }
      }
    }
    return msg;
  },
};

export const GetQuoteResponse = {
  /**
   * Serializes GetQuoteResponse to protobuf.
   */
  encode: function (msg: Partial<GetQuoteResponse>): Uint8Array {
    return GetQuoteResponse._writeMessage(
      msg,
      new BinaryWriter()
    ).getResultBuffer();
  },

  /**
   * Deserializes GetQuoteResponse from protobuf.
   */
  decode: function (bytes: ByteSource): GetQuoteResponse {
    return GetQuoteResponse._readMessage(
      GetQuoteResponse.initialize(),
      new BinaryReader(bytes)
    );
  },

  /**
   * Initializes GetQuoteResponse with all fields set to their default value.
   */
  initialize: function (): GetQuoteResponse {
    return {
      quote: Quote.initialize(),
      metaTransaction: MetaTransaction.initialize(),
    };
  },

  /**
   * @private
   */
  _writeMessage: function (
    msg: Partial<GetQuoteResponse>,
    writer: BinaryWriter
  ): BinaryWriter {
    if (msg.quote) {
      writer.writeMessage(1, msg.quote, Quote._writeMessage);
    }
    if (msg.metaTransaction) {
      writer.writeMessage(
        2,
        msg.metaTransaction,
        MetaTransaction._writeMessage
      );
    }
    return writer;
  },

  /**
   * @private
   */
  _readMessage: function (
    msg: GetQuoteResponse,
    reader: BinaryReader
  ): GetQuoteResponse {
    while (reader.nextField()) {
      const field = reader.getFieldNumber();
      switch (field) {
        case 1: {
          reader.readMessage(msg.quote, Quote._readMessage);
          break;
        }
        case 2: {
          reader.readMessage(msg.metaTransaction, MetaTransaction._readMessage);
          break;
        }
        default: {
          reader.skipField();
          break;
        }
      }
    }
    return msg;
  },
};